refactor: Move dispatchToolCall out of server.ts to tool_dispatch.ts
Verbatim move of the tool-type dispatch switch from a private method on
ActorsMcpServer to a plain exported function, tool_call_error_mapper.ts
style. Both call sites (sync handler, executeToolAndUpdateTask) call the
import directly; server.ts drops 258 lines. No behavior change — #1061
contract suite passes byte-identical. Task machinery deferred, see #1084.

Part of #1084
